--- src/Interfaces/types/post.tsx
@@ -1,9 +1,0 @@
-interface Post {
-  userId: any;
-  id: number;
-  title: string;
-  body: string;
-  comments: string[];
-}
-
-export default Post;
