+++ src/constants/intialForm.ts
@@ -1,0 +1,12 @@
+import { User } from "interfaces/user";
+
+export function initialForm(): User {
+  return {
+    id: null,
+    username: "",
+    email: "",
+    password: "",
+  };
+}
+
+
