+++ src/interfaces/post.ts
@@ -1,10 +1,8 @@
-interface Post {
+export interface Post {
     userId: any;
     id: number;
     title: string;
     body: string;
     comments: string[];
   }
-
-
-export default Post;
+
