+++ src/interfaces/record.ts
