+++ src/interfaces/user.ts
@@ -1,0 +1,6 @@
+export interface User {
+    id: number | null;
+    username: string;
+    email: string;
+    password: string;
+  }
